Renamed base profile files for all eras
