+++ SimPPS/DirectSimProducer/python/profile_2018_postTS2_cff.py
@@ -1,6 +1,6 @@
 import FWCore.ParameterSet.Config as cms
 
-from SimPPS.DirectSimProducer.year_2018_cff import profile_base_2018
+from SimPPS.DirectSimProducer.profile_base_2018_cff import profile_base_2018
 
 alignmentFile = "Validation/CTPPS/alignment/2018_postTS2.xml"
